# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freelancer Platform API
  description: |
    Comprehensive API for a global freelance marketplace with secure payments, 
    verified profiles, milestone-based delivery, and dispute resolution.
  version: 1.0.0
  contact:
    name: NivixPe Team
    email: api@freelancerplatform.com
servers:
  - url: https://api.freelancerplatform.com/v1
    description: Production
  - url: https://staging-api.freelancerplatform.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, OPEN, IN_PROGRESS, COMPLETED, CANCELLED]
        - name: type
          in: query
          schema:
            type: string
            enum: [FIXED_PRICE, HOURLY, CONTEST]
        - name: skills
          in: query
          schema:
            type: array
            items:
              type: string
        - name: budgetMin
          in: query
          schema:
            type: number
        - name: budgetMax
          in: query
          schema:
            type: number
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Projects]
      summary: Create new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get project by ID
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    delete:
      tags: [Projects]
      summary: Delete project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Project deleted
  /projects/{projectId}/bids:
    get:
      tags: [Bids]
      summary: Get bids for project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bid'

  # ==================== BIDS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [FREELANCER, CLIENT, ADMIN]
        fullName:
          type: string
        avatarUrl:
          type: string
        profileData:
          type: object
        rating:
          type: number
        totalReviews:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [FIXED_PRICE, HOURLY, CONTEST]
        status:
          type: string
          enum: [DRAFT, OPEN, IN_PROGRESS, COMPLETED, CANCELLED]
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
          format: date
        visibility:
          type: string
          enum: [PUBLIC, INVITE_ONLY, PRIVATE]
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        client:
          $ref: '#/components/schemas/User'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
    ProjectCreate:
      type: object
      required: [title, description, type, budgetMin, budgetMax]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [FIXED_PRICE, HOURLY, CONTEST]
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
          format: date
        visibility:
          type: string
          enum: [PUBLIC, INVITE_ONLY, PRIVATE]
        skillIds:
          type: array
          items:
            type: string
            format: uuid
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, OPEN, IN_PROGRESS, COMPLETED, CANCELLED]
        budgetMin:
          type: number
        budgetMax:
          type: number
        deadline:
          type: string
          format: date
    Bid:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        freelancerId:
          type: string
          format: uuid
        amount:
          type: number
        deliveryDays:
          type: integer
        proposal:
          type: string
        status:
          type: string
          enum: [PENDING, SHORTLISTED, ACCEPTED, REJECTED, WITHDRAWN]
        createdAt:
          type: string
          format: date-time
        freelancer:
          $ref: '#/components/schemas/User'
...
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'